chore: fix vac cert template to split single qr and multiple qr
